some refact + fitBounds correction
