+++ js/utilities.js
@@ -8,7 +8,15 @@ function getMinOfArray(numArray) {
     return Math.min.apply(null, numArray);
 }
 
+function filterToilets(list, types){
+	return list.filter(function(toilette){
+		return types.indexOf(toilette.class) !== -1 ||
+			((types.indexOf('handicap') !== -1) && toilette.handicap);
+	});
+}
+
 module.exports = {
 	getMinOfArray: getMinOfArray,
-	getMaxOfArray: getMaxOfArray
+	getMaxOfArray: getMaxOfArray,
+	filterToilets: filterToilets
 };
